+++ R/calibrate.R
@@ -55,6 +55,67 @@ calibrateBiomass <- function(params) {
     scaleModel(params, factor = observed_total / model_total)
 }
 
+# The following is a copy of the code for `calibrateBiomass()` just with
+# the text replacements "Biomass" -> "Number" and "biomass" to "number" and
+# the removal of the `params@w` factor in the calculations.
+
+#' Calibrate the model scale to match total observed number
+#'
+#' `r lifecycle::badge("experimental")`
+#' Given a MizerParams object `params` for which number observations are
+#' available for at least some species via the `number_observed` column in the
+#' species_params data frame, this function returns an updated MizerParams
+#' object which is rescaled with [scaleModel()] so that the total number in
+#' the model agrees with the total observed number.
+#'
+#' Number observations usually only include individuals above a certain size.
+#' This size should be specified in a number_cutoff column of the species
+#' parameter data frame. If this is missing, it is assumed that all sizes are
+#' included in the observed number, i.e., it includes larval number.
+#'
+#' After using this function the total number in the model will match the
+#' total number, summed over all species. However the numbers of the
+#' individual species will not match observations yet, with some species
+#' having numbers that are too high and others too low. So after this
+#' function you may want to use [matchNumbers()]. This is described in the
+#' blog post at https://bit.ly/2YqXESV.
+#'
+#' If you have observations of the yearly yield instead of numbers, you can
+#' use [calibrateYield()] instead of this function.
+#'
+#' @param params A MizerParams object
+#' @return A MizerParams object
+#' @export
+#' @examples
+#' params <- NS_params
+#' species_params(params)$number_observed <-
+#'     c(0.8, 61, 12, 35, 1.6, 20, 10, 7.6, 135, 60, 30, 78)
+#' species_params(params)$number_cutoff <- 10
+#' params2 <- calibrateNumber(params)
+#' plotNumberVsSpecies(params2)
+calibrateNumber <- function(params) {
+    if ((!("number_observed" %in% names(params@species_params))) ||
+        all(is.na(params@species_params$number_observed))) {
+        return(params)
+    }
+    no_sp <- nrow(params@species_params)
+    cutoff <- params@species_params$number_cutoff
+    # When no cutoff known, set it to 0
+    if (is.null(cutoff)) cutoff <- rep(0, no_sp)
+    cutoff[is.na(cutoff)] <- 0
+    observed <- params@species_params$number_observed
+    observed_total <- sum(observed, na.rm = TRUE)
+    sp_observed <- which(!is.na(observed))
+    model_total <- 0
+    for (sp_idx in sp_observed) {
+        model_total <-
+            model_total +
+            sum((params@initial_n[sp_idx, ] * params@dw)
+                [params@w >= cutoff[[sp_idx]]])
+    }
+    scaleModel(params, factor = observed_total / model_total)
+}
+
 #' Calibrate the model scale to match total observed yield
 #' 
 #' `r lifecycle::badge("experimental")`